move fix-instructions-panel to common
