+++ src/tests/unit/tests/injected/components/fix-instruction-panel.test.tsx
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { FixInstructionPanel, FixInstructionPanelProps } from 'common/components/fix-instruction-panel';
+import { FixInstructionProcessor } from 'common/components/fix-instruction-processor';
 import { shallow } from 'enzyme';
 import { CheckType } from 'injected/components/details-dialog';
-import { FixInstructionProcessor } from 'injected/fix-instruction-processor';
 import * as React from 'react';
 import { IMock, It, Mock, MockBehavior } from 'typemoq';
